==> 04-rossman/kaggle-rossmann-master/04-Normalize-and_encode.ipynb <==
! pip install pandas_summary

import pandas as pd
import numpy as np
from pandas_summary import DataFrameSummary
from sklearn.preprocessing import LabelEncoder, StandardScaler

from google.colab import drive
drive.mount('/content/drive')

# pip install pyarrow

PATH = '/content/drive/My Drive/Colab Notebooks/kaggle-rossmann-master/rossmann/rossmann/'
df_train = pd.read_feather(PATH+'all_preprocessed_train.fth')
df_test = pd.read_feather(PATH+'all_preprocessed_test.fth')

df_train.shape, df_test.shape

# Verificación de que se hicieron bien las cosas en testa también
set(df_train.columns) - set(df_test.columns)

(df_train['CompetitionDistance'].isna()).sum()

## Categorical vars

cat_vars = ['Store', 'DayOfWeek', 'Year', 'Month', 'Day', 'StateHoliday', 
            'CompetitionMonthsOpen', 'Promo2Weeks', 
            'StoreType', 'Assortment', 'PromoInterval', 
            'CompetitionOpenSinceYear', 'Promo2SinceYear', 
            'State', 'Week', 'Events', 'Promo_fw', 'Promo_bw', 
            'StateHoliday_bool_fw', 'StateHoliday_bool_bw', 'SchoolHoliday_fw', 
            'SchoolHoliday_bw']

#DataFrameSummary(df_train[cat_vars]).summary()

DataFrameSummary(df_train[cat_vars]).summary().loc[['uniques', 'types', 'missing']].T

# Reemplaza nulls por "" en categoricas
# for v in cat_vars: df_train.loc[df_train[v].isnull(), v] = ""

# for v in cat_vars: df_test.loc[df_test[v].isnull(), v] = ""

# Continuos Vars

df_train.columns

contin_vars = ['CompetitionDistance', 
               'Max_TemperatureC', 'Mean_TemperatureC', 'Min_TemperatureC', 
               'Precipitationmm', 'Max_Humidity', 'Mean_Humidity', 'Min_Humidity', 
               'Max_Wind_SpeedKm_h', 'Mean_Wind_SpeedKm_h', 'CloudCover', 'trend', 
               'trend_DE', 'AfterStateHoliday_bool', 'BeforeStateHoliday_bool', 
               'Promo', 'SchoolHoliday', 'StateHoliday_bool']

DataFrameSummary(df_train[contin_vars]).summary().loc[['uniques', 'types', 'missing']].T

# Reemplaza nulls por 0 en continuas
# for v in contin_vars: df_train.loc[df_train[v].isnull(), v] = 0
# for v in contin_vars: df_test.loc[df_test[v].isnull(), v] = 0

# Normalización en continuas y LabelEncode en Categoricas

# pip install sklearn_pandas
from sklearn_pandas import DataFrameMapper

cat_maps = [(o, LabelEncoder()) for o in cat_vars]
contin_maps = [([o], StandardScaler()) for o in contin_vars]

mapper_cat = DataFrameMapper(cat_maps)
_ = mapper_cat.fit(df_train)

# Assortment es la novena posición en cat_maps
N = 10
print(list(zip(df_train['Assortment'].values[:N], mapper_cat.transform(df_train)[:,9][:N])))
print(list(zip(df_train['Events'].values[:N], mapper_cat.transform(df_train)[:,15][:N])))
print(list(zip(df_train['Year'].values[:N], mapper_cat.transform(df_train)[:,2][:N])))

mapper_cont = DataFrameMapper(contin_maps)
_ = mapper_cont.fit(df_train)

N = 10
print(df_train['CompetitionDistance'].values[:N])
print(mapper_cont.transform(df_train)[:, 0][:N])

# La hacemos con todas ahora
df_train[cat_vars] = mapper_cat.transform(df_train)
df_test[cat_vars] = mapper_cat.transform(df_test)

df_train[contin_vars] = mapper_cont.transform(df_train)
df_test[contin_vars] = mapper_cont.transform(df_test)

df_train.head()

DataFrameSummary(df_train[cat_vars]).summary().loc[['uniques', 'types']]

# Final processing

df_train.shape

df_sales = df_train[df_train.Sales!=0]

df_sales.shape

df_test.reset_index(inplace=True)

df_sales.reset_index(inplace=True)

df_sales.to_feather('train_normalized_data.fth')

df_test.to_feather('test_normalized_data.fth')

(df_test['Open'] == 0).sum()

(df_sales['Open'] == 0).sum()

(df_train['CompetitionDistance'].isna()).sum()
